parser/tests/tree_construction/basic.dat

#data
<a><b><div id=1><div id=2><div id=3><div id=4><div id=5><div id=6><div id=7><div id=8>A</a>
#errors
(1,3): expected-doctype-but-got-start-tag
(1,91): adoption-agency-1.3
(1,91): adoption-agency-1.3
(1,91): adoption-agency-1.3
(1,91): adoption-agency-1.3
(1,91): adoption-agency-1.3
(1,91): adoption-agency-1.3
(1,91): adoption-agency-1.3
(1,91): adoption-agency-1.3
(1,91): expected-closing-tag-but-got-eof
#document
| <html>
|   <head>
|   <body>
|     <a>
|       <b>
|     <b>
|       <div>
|         id="1"
|         <a>
|         <div>
|           id="2"
|           <a>
|           <div>
|             id="3"
|             <a>
|             <div>
|               id="4"
|               <a>
|               <div>
|                 id="5"
|                 <a>
|                 <div>
|                   id="6"
|                   <a>
|                   <div>
|                     id="7"
|                     <a>
|                     <div>
|                       id="8"
|                       <a>
|                         "A"
